% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPostEst.R
\docType{package}
\name{BayesPostEst}
\alias{BayesPostEst}
\alias{BayesPostEst-package}
\title{BayesPostEst}
\description{
BayesPostEst
}
\section{BayesPostEst functions}{
 The functions in this
package assist in analyzing and presenting results
from Bayesian analysis.
}

