% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bugs_logit}
\alias{bugs_logit}
\title{Fitted BUGS logit model}
\format{
a R2WinBUGS bugs object.
}
\usage{
bugs_logit
}
\description{
A fitted WinBUGS logit model.
}
\examples{
data(bugs_logit)
bugs_logit  
}
\keyword{datasets}
