% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bugs_model}
\alias{bugs_model}
\title{Fitted BUGS model}
\format{
A class "bugs" object created by [R2WinBUGS::as.bugs.array()]
}
\usage{
bugs_model
}
\description{
A fitted BUGS model with generated with
[R2WinBUGS::as.bugs.array()]. See the example  code below for how it was created. Used
in examples and for testing.
}
\examples{
\donttest{
data(LINE, package = "rjags")
LINE$recompile()

## fitting the model with jags
bugs_model <- rjags::coda.samples(LINE, c("alpha", "beta", "sigma"),
                                  n.iter = 1000)

bugs_model <- R2WinBUGS::as.bugs.array(sims.array = as.array(bugs_model))
}
}
\keyword{datasets}
