% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRocPrc.R
\name{identify_link_function}
\alias{identify_link_function}
\title{Try to identify link function}
\usage{
identify_link_function(obj)
}
\arguments{
\item{obj}{stanfit object}
}
\value{
Either "logit" or "probit"; if neither can be identified the function
will return `NA_character_`.
}
\description{
Try to identify link function
}
\keyword{internal}
