% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRocPrc.R
\name{is_binary_model}
\alias{is_binary_model}
\title{Try to identify if a stanfit model is a binary choice model}
\usage{
is_binary_model(obj)
}
\arguments{
\item{obj}{stanfit object}
}
\description{
Try to identify if a stanfit model is a binary choice model
}
\keyword{internal}
