% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jags_interactive}
\alias{jags_interactive}
\title{Fitted JAGS interactive linear model}
\format{A class "rjags" object created by [R2jags::jags()]}
\usage{
jags_interactive
}
\description{
A fitted JAGS linear model with interaction term generated with
[R2jags::jags()]. See the example  code below for how it was created. Used
in examples and for testing.
}
\keyword{datasets}
