% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_roc_prc.R
\name{MCMC_roc_prc}
\alias{MCMC_roc_prc}
\title{MCMC ROC PRC}
\usage{
MCMC_roc_prc(stan_object, model_frame, model_name = "Model")
}
\arguments{
\item{stan_object}{rstan object of class "stanfit", including parameters and 
predicted values for each observation, with the prefix "pred"}

\item{model_frame}{model frame used to fit the model}

\item{model_name}{name of model for plotting}
}
\value{
A list with four elements: 
area_under_roc: area under ROC curve (scalar)
area_under_prc: area under precision-recall curve (scalar)
prc_dat: data to plot precision-recall curve (data frame)
roc_dat: data to plot ROC curve (data frame)
}
\description{
This function generates ROC and precision-recall curves 
after fitting a logit/probit regression model in Stan
references: Area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\details{
This function generates ROC and precision-recall curves 
after fitting a logit/probit regression model in Stan
references: Area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
