% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcObsProb.R
\name{mcmcObsProb}
\alias{mcmcObsProb}
\title{Bayesian MCMC Observed Values Predicted Probablities}
\usage{
mcmcObsProb(model_matrix, mcmc_out, xcol, xrange, link = "logit",
  lower = 0.05, upper = 0.95)
}
\arguments{
\item{model_matrix}{model matrix, including intercept. Create with 
model.matrix(formula, data)}

\item{mcmc_out}{posterior distributions of all coefficients in matrix 
form - can easily be created from rstan, MCMCpack, R2jags, R2OpenBUGS, etc.}

\item{xcol}{column number of the explanatory variable for which to calculate 
associated Pr(y = 1)}

\item{xrange}{name of the vector with the range of relevant values of the 
explanatory variable for which to calculate associated Pr(y = 1)}

\item{link}{link function, character vector set to "logit" (default) or "probit"}

\item{lower}{lower bound of credible interval of predicted probability at given x}

\item{upper}{upper bound of credible interval of predicted probability at given x}
}
\value{
An object
}
\description{
Implements R function to calculate the predicted probabilities
for "observed" cases after a Bayesian logit or probit model, following
Hanmer & Kalkan (2013; doi: 10.1111/j.1540-5907.2012.00602.x).

Please check out and cite:

Hanmer & Kalkan (2013; doi: 10.1111/j.1540-5907.2012.00602.x).
}
\details{
This function implements calculates predicted probabilities for 
"observed" cases after a Bayesian logit or probit model,
following Hanmer & Kalkan (2013; doi: 10.1111/j.1540-5907.2012.00602.x).
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
