% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmcpack_linear}
\alias{mcmcpack_linear}
\title{Fitted MCMCpack  linear model}
\format{
A class "mcmc" object created by [MCMCpack::MCMCregress()]
}
\usage{
mcmcpack_linear
}
\description{
A fitted MCMCpack linear model with interaction term generated with
[MCMCpack::MCMCregress()]. See the example code below for how it was created. Used
in examples and for testing.
}
\examples{
\donttest{
## fitting the model with MCMCpack
mcmcpack_linear <- MCMCpack::MCMCregress(Y ~ X, b0 = 0, B0 = 0.001,
                                         sigma.mu = 5, sigma.var = 10,
                                         data = list(X = rnorm(100),
                                                     Y = rnorm(100, 5, 5)),
                                         seed = 1)
}
}
\keyword{datasets}
