% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCprobit.pp.plot.R
\name{MCMCprobit.pp.plot}
\alias{MCMCprobit.pp.plot}
\title{MCMCprobit.pp.plot}
\usage{
MCMCprobit.pp.plot(model.matrix, mcmc.out, x.col = 2, x.range, xlabel,
  ylabel)
}
\arguments{
\item{x.col}{tbd}

\item{x.range}{tbd}

\item{xlabel}{tbd}

\item{ylabel}{tbd}

\item{model_matrix}{model matrix, including intercept, focal predictor in the second column}

\item{mcmc_out}{posterior distributions of all probit coefficients, 
in matrix form - can easily be created from rstan, MCMCpack, R2jags, etc.}
}
\value{
output
}
\description{
R function to calculate and plot predicted probabilities after a Bayesian probit model
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
