% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRocPrc.R
\name{new_mcmcRocPrc}
\alias{new_mcmcRocPrc}
\title{Constructor for mcmcRocPrc objects}
\usage{
new_mcmcRocPrc(pred_prob, yvec, curves, fullsims)
}
\arguments{
\item{pred_prob}{a \verb{\[N, iter\]} matrix of predicted probabilities}

\item{yvec}{a \code{numeric(N)} vector of observed outcomes}

\item{curves}{include curve data in output?}

\item{fullsims}{collapse posterior samples into single summary?}
}
\description{
This function actually does the heavy lifting once we have a matrix of
predicted probabilities from a model, plus the vector of observed outcomes.
The reason to have it here in a single function is that we don't replicate
it in each function that accomodates a JAGS, BUGS, RStan, etc. object.
}
\keyword{internal}
