% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runjags_interactive}
\alias{runjags_interactive}
\title{Fitted runjags interactive linear model}
\format{
A class "runjags" object created by [runjags::run.jags()]
}
\usage{
runjags_interactive
}
\description{
A fitted JAGS linear model with interaction term generated with
[runjags::run.jags()]. See the example  code below for how it was created. Used
in examples and for testing.
}
\examples{
\donttest{
library(runjags)

data("sim_data_interactive")

## formatting the data for jags
datalist <- list(X = model.matrix(~ X1 * X2, sim_data_interactive),
                 Y = sim_data_interactive[, 3],
                 N = nrow(sim_data_interactive))

## creating jags model
model <- "model { 
for(i in 1 : N){ 
	Y[i] ~ dnorm(beta \%*\% X[i, ], tau);
} 
for(i in 1:4) {
  beta[i] ~ dnorm(0, 0.001)
}
tau ~ dexp(1)
}"

## fitting the model with runjags
runjags_interactive <- run.jags(model = model, monitor = c("beta", "tau"),
                                data = datalist, n.chains = 2, method = "rjags",
                                inits = list(list(beta = rnorm(4)), 
                                             list(beta = rnorm(4)))
                                )
}
}
\keyword{datasets}
