% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPostEst.R
\docType{package}
\name{BayesPostEst}
\alias{BayesPostEst}
\alias{BayesPostEst-package}
\title{BayesPostEst}
\description{
BayesPostEst
}
\section{BayesPostEst functions}{
 This package currently has 
seven main functions 
\itemize{
\item \code{mcmcAveProb()}
\item \code{mcmcObsProb()}
\item \code{mcmcFD()}
\item \code{mcmcFDplot()}
\item \code{mcmcRocPrc()}
\item \code{mcmcTab()}
\item \code{mcmcReg()}
}
These functions can be used to generate and plot postestimation quantities after 
estimating Bayesian regression models using MCMC. The package combines functions 
written originally for Johannes Karreth's workshop on Bayesian modeling at the 
ICPSR Summer program. For now, the package focuses mostly on generalized linear 
regression models for binary outcomes (logistic and probit regression). 
The vignette for this package has a walk-through of each function in action. 
Please refer to that to get an overview of all the functions, or visit the 
documentation for a specific function of your choice. Johannes Karreth's website
(http://www.jkarreth.net) also has resources for getting started with Bayesian 
analysis, fitting models, and presenting results.
}

