% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPostEst.R
\docType{package}
\name{BayesPostEst}
\alias{BayesPostEst}
\alias{BayesPostEst-package}
\title{BayesPostEst}
\description{
BayesPostEst
}
\section{BayesPostEst functions}{

To use this package, decide which function you need to use
to present and visualize your Bayesian results.
}

