% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPostEst-package.R
\docType{package}
\name{BayesPostEst}
\alias{BayesPostEst}
\alias{BayesPostEst-package}
\title{BayesPostEst Overview}
\description{
This package currently has nine main functions that can be used to generate 
and plot postestimation quantities after estimating Bayesian regression models using MCMC. 
The package combines functions written originally for Johannes Karreth's workshop on 
Bayesian modeling at the ICPSR Summer program. Currently BayesPostEst focuses mostly on 
generalized linear regression models for binary outcomes (logistic and probit regression). 
The vignette for this package has a walk-through of each function in action. 
Please refer to that to get an overview of all the functions, or visit the 
documentation for a specific function of your choice. Johannes Karreth's website
(http://www.jkarreth.net) also has resources for getting started with Bayesian 
analysis, fitting models, and presenting results.
}
\section{Main Functions}{

\itemize{
\item \code{mcmcAveProb()}
\item \code{mcmcObsProb()}
\item \code{mcmcFD()}
\item \code{mcmcMargEff()}
\item \code{mcmcRocPrc()}
\item \code{mcmcRocPrcGen()}
\item \code{mcmcTab()}
\item \code{mcmcReg()}
\item \code{plot.mcmcFD()}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShanaScogin/BayesPostEst}
  \item Report bugs at \url{https://github.com/ShanaScogin/BayesPostEst/issues}
}

}
\author{
\strong{Maintainer}: Shana Scogin \email{shanarscogin@gmail.com} (\href{https://orcid.org/0000-0002-7801-853X}{ORCID})

Authors:
\itemize{
  \item Johannes Karreth \email{jkarreth@ursinus.edu} (\href{https://orcid.org/0000-0003-4586-7153}{ORCID})
  \item Rob Williams \email{jayrobwilliams@gmail.com} (\href{https://orcid.org/0000-0001-9259-3883}{ORCID})
  \item Andreas Beger \email{adbeger@gmail.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})
}

Other contributors:
\itemize{
  \item Myunghee Lee [contributor]
  \item Neil Williams [contributor]
}

}
