% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfBarTable.R
\name{dfBarTable}
\alias{dfBarTable}
\title{Generate bar plots for multiple variables}
\usage{
dfBarTable(df, rows = 2)
}
\arguments{
\item{rows}{number of rows of plots within the figure}
}
\value{
output
}
\description{
R function for summarizing a data frame as a table of bar plots
}
\details{
R function for summarizing a data frame as a table of bar plots
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
