% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcAveProb.R
\name{mcmcAveProb}
\alias{mcmcAveProb}
\title{Bayesian MCMC Predicted Probablities for the 'Average' Case}
\usage{
mcmcAveProb(modelmatrix, mcmcout, xcol, xrange, xinterest,
  link = "logit", ci = c(0.025, 0.975), fullsims = FALSE)
}
\arguments{
\item{modelmatrix}{model matrix, including intercept. Create with 
\code{model.matrix(formula, data)}}

\item{mcmcout}{posterior distributions of all coefficients
in matrix form - can easily be created from rstan, MCMCpack, R2jags, R2OpenBUGS, etc.}

\item{xcol}{column number of the explanatory variable for which to calculate 
associated Pr(y = 1)}

\item{xrange}{name of the vector with the range of relevant values of the 
explanatory variable for which to calculate associated Pr(y = 1)}

\item{xinterest}{name of the explanatory variable for which to calculate 
associated Pr(y = 1). If xcol is supplied, this is not needed. If both
are supplied, the function defaults to this argument and xcol is ignored}

\item{link}{type of model. It is a character vector set to 
"logit" (default) or "probit"}

\item{ci}{the bounds of the credible interval. Default is \code{c(0.05, 0.95)}.}

\item{fullsims}{logical indicator of whether full object (based on all MCMC draws 
rather than average) will be returned. Default is \code{FALSE}. A note: The longer 
\code{xrange} is, the larger the full output will be if \code{TRUE} is selected}
}
\value{
This function returns a matrix with 4 columns:
\itemize{
\item predictor: identical to x_range
\item median_pp: median predicted probability at given x
\item lower_pp: lower bound of credible interval of predicted probability at given x
\item upper_pp: upper bound of credible interval of predicted probability at given x
}
}
\description{
This function calculates predicted probabilities for "average" cases after 
a Bayesian logit or probit model. For an explanation of predicted probabilities for 
"average" cases, see e.g. King, Tomz & Wittenberg (2000) <doi: 10.2307/2669316>
}
\details{
This function calculates predicted probabilities for "average" cases after a Bayesian 
logit or probit model. For an explanation of predicted probabilities for "average" cases,
see e.g. King, Tomz & Wittenberg (2000) <doi: 10.2307/2669316>
}
\examples{
\donttest{
  ## simulating data
  set.seed(123456)
  b0 <- 0.2 # true value for the intercept
  b1 <- 0.5 # true value for first beta
  b2 <- 0.7 # true value for second beta
  n <- 500 # sample size
  X1 <- runif(n, -1, 1)
  X2 <- runif(n, -1, 1)
  Z <- b0 + b1 * X1 + b2 * X2
  pr <- 1 / (1 + exp(-Z)) # inv logit function
  Y <- rbinom(n, 1, pr) 
  data <- data.frame(cbind(X1, X2, Y))
  
  ## formatting the data for jags
  datjags <- as.list(data)
  datjags$N <- length(datjags$Y)
  
  ## creating jags model
  model <- function()  {
  
  for(i in 1:N){
    Y[i] ~ dbern(p[i])  ## Bernoulli distribution of y_i
    logit(p[i]) <- mu[i]    ## Logit link function
    mu[i] <- b[1] + 
      b[2] * X1[i] + 
      b[3] * X2[i]
  }
  
  for(j in 1:3){
    b[j] ~ dnorm(0, 0.001) ## Use a coefficient vector for simplicity
  }
  
}

params <- c("b")
inits1 <- list("b" = rep(0, 3))
inits2 <- list("b" = rep(0, 3))
inits <- list(inits1, inits2)

## fitting the model with R2jags
library(R2jags)
set.seed(123)
fit <- jags(data = datjags, inits = inits, 
         parameters.to.save = params, n.chains = 2, n.iter = 2000, 
         n.burnin = 1000, model.file = model)

### average value approach
library(coda)
xmat <- model.matrix(Y ~ X1 + X2, data = data)
mcmc <- as.mcmc(fit)
mcmc_mat <- as.matrix(mcmc)[, 1:ncol(xmat)]
X1_sim <- seq(from = min(datjags$X1),
              to = max(datjags$X1), 
              length.out = 10)
ave_prob <- mcmcAveProb(modelmatrix = xmat,
                        mcmcout = mcmc_mat,
                        xrange = X1_sim, 
                        xcol = 2)
}
}
\references{
King, G., Tomz, M., & Wittenberg, J. (2000). Making the most of 
statistical analyses: Improving interpretation and presentation. Available at 
SSRN 1083738. <doi: 10.2307/2669316>
}
