% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcAveProb.R
\name{mcmcAveProb}
\alias{mcmcAveProb}
\title{Bayesian MCMC Predicted Probablities for the 'Average' Case}
\usage{
mcmcAveProb(model_matrix, mcmc_out, x_col, x_range_vec, link = "logit",
  ci = c(0.05, 0.95))
}
\arguments{
\item{model_matrix}{model matrix, including intercept. Create with model.matrix(formula, data)}

\item{mcmc_out}{posterior distributions of all coefficients}

\item{x_col}{column number of the explanatory variable for which to calculate 
associated Pr(y = 1)}

\item{x_range_vec}{name of the vector with the range of relevant values of the 
explanatory variable for which to calculate associated Pr(y = 1)}

\item{link}{link: link function, character vector set to "logit" (default) or "probit"}

\item{ci}{the bounds of the credible interval. Default is 0.05 and 0.95. 
Enter as a vector, such as c(0.05, 0.95).
#'@return a matrix with 4 columns:
predictor: identical to x_range
median_pp: median predicted probability at given x
lower_pp: lower bound of credible interval of predicted probability at given x
upper_pp: upper bound of credible interval of predicted probability at given x}
}
\description{
This function calculates predicted probabilities for "average" cases after a Bayesian logit 
or probit model. For an explanation of predicted probabilities for "average" cases,
}
\details{
This function calculates predicted probabilities for "average" cases after a Bayesian 
logit or probit model. For an explanation of predicted probabilities for "average" cases,
see e.g. King, Tomz & Wittenberg (2000, <doi: 10.2307/2669316>)
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
