% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFD.R
\name{mcmcFD}
\alias{mcmcFD}
\title{First Differences of a Bayesian Logit or Probit model}
\usage{
mcmcFD(modelmatrix, mcmcout, link = "logit", ci = c(0.025, 0.975),
  percentiles = c(0.25, 0.75), fullsims = FALSE)
}
\arguments{
\item{modelmatrix}{model matrix, including intercept. Create with model.matrix(formula, data)}

\item{mcmcout}{posterior distributions of all logit coefficients, 
in matrix form - can easily be created from rstan, MCMCpack, R2jags, etc.}

\item{link}{type of model. It is a character vector set to 
"logit" (default) or "probit"}

\item{ci}{the bounds of the credible interval. Default is \code{c(0.025, 0.975)}.}

\item{percentiles}{default is \code{c(0.25, 0.75)}}

\item{fullsims}{logical indicator of whether full object (based on all MCMC draws 
rather than average) will be returned. Default is \code{FALSE}}
}
\value{
output
}
\description{
R function to calculate first differences after a Bayesian logit or probit model
}
\examples{
\donttest{
  set.seed(123456)
}
}
