% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFD.R
\name{mcmcFD}
\alias{mcmcFD}
\title{First Differences of a Bayesian Logit or Probit model}
\usage{
mcmcFD(model_matrix, mcmc_out, link = "logit", ci = c(0.05, 0.95),
  percentiles = c(0.25, 0.75), full_sims = FALSE)
}
\arguments{
\item{model_matrix}{model matrix, including intercept. Create with model.matrix(formula, data)}

\item{mcmc_out}{posterior distributions of all logit coefficients, 
in matrix form - can easily be created from rstan, MCMCpack, R2jags, etc.}

\item{ci}{the bounds of the credible interval}

\item{percentiles}{tbd}

\item{full_sims}{tbd}
}
\value{
output
}
\description{
R function to calculate first differences after a Bayesian logit or probit model
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
