% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFDplot.R
\name{mcmcFDPlot}
\alias{mcmcFDPlot}
\title{Plot Predicted Probabilities after a Bayesian Probit Model}
\usage{
mcmcFDPlot(model.matrix, mcmc.out, xcol = 2, xrange, xlabel, ylabel)
}
\arguments{
\item{xcol}{tbd}

\item{xrange}{tbd}

\item{xlabel}{tbd}

\item{ylabel}{tbd}

\item{model_matrix}{model matrix, including intercept, focal predictor in the second column}

\item{mcmc_out}{posterior distributions of all probit coefficients, 
in matrix form - can easily be created from rstan, MCMCpack, R2jags, etc.}
}
\value{
output
}
\description{
R function to calculate and plot predicted probabilities after a Bayesian probit model
}
\details{
An R function to calculate and plot predicted probabilities after a Bayesian probit model.
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}

}
