% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFDplot.R
\name{mcmcFDplot}
\alias{mcmcFDplot}
\title{Plot First Differences from MCMC output}
\usage{
mcmcFDplot(fdfull, ROPE = NULL)
}
\arguments{
\item{fdfull}{Output generated from mcmcFD(..., full_sims = TRUE)}

\item{ROPE}{numeric vector of length two, defining the Region of Practical 
Equivalence around 0. Defaults to NULL}
}
\value{
output
}
\description{
R function to plot first differences generated from MCMC output
}
\details{
An R function to plot first differences after a Bayesian logit or probit model.
}
\examples{
\donttest{
  set.seed(123456)
}
}
