% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRocPrc.R
\name{mcmcRocPrc}
\alias{mcmcRocPrc}
\title{ROC and Precision-Recall Curves using Bayesian MCMC estimates}
\usage{
mcmcRocPrc(modelmatrix, mcmcout, modelframe, curves = FALSE,
  link = "logit", fullsims = FALSE)
}
\arguments{
\item{modelmatrix}{model matrix, including intercept (if the intercept is among the
parameters estimated in the model). Create with model.matrix(formula, data).
Note: the order of columns in the model matrix must correspond to the order of columns 
in the matrix of posterior draws in the \code{mcmcout} argument. See the \code{mcmcout}
argument for more.}

\item{mcmcout}{posterior distributions of all logit coefficients, 
in matrix form. This can be created from rstan, MCMCpack, R2jags, etc. and transformed
into a matrix using the function as.mcmc() from the coda package for \code{jags} class
objects, as.matrix() from base R for \code{mcmc}, \code{mcmc.list}, \code{stanreg}, and 
\code{stanfit} class objects, and \code{object$sims.matrix} for \code{bugs} class objects.
Note: the order of columns in this matrix must correspond to the order of columns 
in the model matrix. One can do this by examining the posterior distribution matrix and sorting the 
variables in the order of this matrix when creating the model matrix. A useful function for sorting 
column names containing both characters and numbers as 
you create the matrix of posterior distributions is \code{mixedsort()} fom the gtools package.}

\item{modelframe}{model frame in matrix form. Can be created using 
as.matrix(model.frame(formula, data))}

\item{curves}{logical indicator of whether or not to return values to plot the ROC or Precision-Recall 
curves. If set to \code{FALSE} (default), results are returned as a list without the extra
values.}

\item{link}{type of generalized linear model; a character vector set to \code{"logit"} (default) or \code{"probit"}.}

\item{fullsims}{logical indicator of whether full object (based on all MCMC draws 
rather than their average) will be returned. Default is \code{FALSE}. Note:  If \code{TRUE}
is chosen, the function takes notably longer to execute.}
}
\value{
This function returns a list with 4 elements:
\itemize{
\item area_under_roc: area under ROC curve (scalar)
\item area_under_prc: area under precision-recall curve (scalar)
\item prc_dat: data to plot precision-recall curve (data frame)
\item roc_dat: data to plot ROC curve (data frame)
}
}
\description{
This function generates ROC and Precision-Recall curves 
after fitting a Bayesian logit or probit regression . For more,
please see other area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>.
}
\details{
This function generates ROC and precision-recall curves 
after fitting a Bayesian logit or probit model. For more,
please see other area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\examples{
\donttest{
# simulating data
set.seed(123456)
b0 <- 0.2 # true value for the intercept
b1 <- 0.5 # true value for first beta
b2 <- 0.7 # true value for second beta
n <- 500 # sample size
X1 <- runif(n, -1, 1)
X2 <- runif(n, -1, 1)
Z <- b0 + b1 * X1 + b2 * X2
pr <- 1 / (1 + exp(-Z)) # inv logit function
Y <- rbinom(n, 1, pr) 
data <- data.frame(cbind(X1, X2, Y))

## fitting with stan
library("rstanarm")
m <- rstanarm::stan_glm(Y ~ X1 + X2,
                        data = data,
                        family = binomial(link = "logit"))

## using mcmcRocPrc with median draws
fit_sum <- mcmcRocPrc(sims = m,
                      modelframe = model.frame(m),
                      fullsims = FALSE)

fit_sum$area_under_roc
plot(x = fit_sum$roc_dat$x, 
     y = fit_sum$roc_dat$y, 
     type = "l", 
     main = "ROC")
abline(a = 0, b = 1)

fit_sum$area_under_prc
plot(x = fit_sum$prc_dat$x, 
     y = fit_sum$prc_dat$y, 
     type = "l", 
     main = "PRC")

}
}
\references{
Beger, Andreas. 2016. “Precision-Recall Curves.” Available at SSRN: 
https://ssrn.com/Abstract=2765419. http://dx.doi.org/10.2139/ssrn.2765419.
}
