% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRocPrc.R
\name{mcmcRocPrc}
\alias{mcmcRocPrc}
\title{ROC and Precision-recall Curve}
\usage{
mcmcRocPrc(sims, modelframe, fullsims = FALSE)
}
\arguments{
\item{sims}{MCMC object generated by R2jags, rjags, R2WinBUGS, R2OpenBUGS, 
MCMCpack, rstan, rstanarm. Currently only rstanarm objects are supported}

\item{modelframe}{model frame used to fit the model}

\item{fullsims}{logical indicator of whether full object (based on all MCMC draws 
rather than average) will be returned. Default is \code{FALSE}}
}
\value{
This function returns a list with 4 elements:
\itemize{
\item area_under_roc: area under ROC curve (scalar)
\item area_under_prc: area under precision-recall curve (scalar)
\item prc_dat: data to plot precision-recall curve (data frame)
\item roc_dat: data to plot ROC curve (data frame)
}
}
\description{
This function generates ROC and precision-recall curves 
after fitting a logit/probit regression model in Stan
references: Area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\details{
This function generates ROC and precision-recall curves 
after fitting a logit/probit regression model in Stan. For more,
please see other area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\examples{
\donttest{
  set.seed(123456)
}
}
\references{
Beger, A. (2016). Precision-recall curves.
<https://ssrn.com/abstract=2765419>
}
