% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRocPrc.R
\name{mcmcRocPrc}
\alias{mcmcRocPrc}
\title{ROC and Precision-recall Curve}
\usage{
mcmcRocPrc(sims, model_frame, model_name = "Model")
}
\arguments{
\item{sims}{MCMC object generated by R2jags, rjags, R2WinBUGS, R2OpenBUGS, 
MCMCpack, rstan, rstanarm}

\item{model_frame}{model frame used to fit the model}

\item{model_name}{name of model for plotting}
}
\value{
A list with four elements: 
area_under_roc: area under ROC curve (scalar)
area_under_prc: area under precision-recall curve (scalar)
prc_dat: data to plot precision-recall curve (data frame)
roc_dat: data to plot ROC curve (data frame)
}
\description{
This function generates ROC and precision-recall curves 
after fitting a logit/probit regression model in Stan
references: Area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\details{
This function generates ROC and precision-recall curves 
after fitting a logit/probit regression model in Stan
references: Area under the curve functions written by Andreas Beger
<https://ssrn.com/abstract=2765419>
}
\examples{
\donttest{
  set.seed(123456)
}
}
