% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTab.R
\name{mcmcTab}
\alias{mcmcTab}
\title{Summarize Bayesian MCMC Output}
\usage{
mcmcTab(sims, ci = 0.95, pars = NULL, Pr = FALSE)
}
\arguments{
\item{sims}{MCMC object generated by R2jags, rjags, R2WinBUGS, R2OpenBUGS, MCMCpack, rstan, rstanarm}

\item{ci}{desired level for credible intervals; defaults to 0.95}

\item{pars}{character vector of parameters to be printed; defaults to NULL (all parameters are printed)}

\item{Pr}{print percent of posterior draws with same sign as median; defaults to TRUE}
}
\value{
output
}
\description{
R function for summarizing MCMC output in a regression-style table
}
\details{
R function for summarizing MCMC output in a regression-style table
}
\examples{
\donttest{
  set.seed(123456)
  example for user goes here
  unit testing goes in testthat
}
}
