% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTab.R
\name{mcmcTab}
\alias{mcmcTab}
\title{Summarize Bayesian MCMC Output}
\usage{
mcmcTab(sims, ci = c(0.025, 0.975), pars = NULL, Pr = FALSE,
  ROPE = NULL)
}
\arguments{
\item{sims}{MCMC object generated by R2jags, rjags, R2WinBUGS, R2OpenBUGS, 
MCMCpack, rstan, rstanarm}

\item{ci}{desired level for credible intervals; defaults to c(0.025, 0.975)}

\item{pars}{character vector of parameters to be printed; defaults to NULL (all parameters are printed)}

\item{Pr}{print percent of posterior draws with same sign as median; defaults to FALSE}

\item{ROPE}{vector of two values defining the region of practical equivalence; 
returns \% of posterior to the left/right of ROPE}
}
\value{
an data frame containing MCMC summary statistics
}
\description{
R function for summarizing MCMC output in a regression-style table
}
\details{
R function for summarizing MCMC output in a regression-style table
}
\examples{
\donttest{
## simulating data
set.seed(123456)
b0 <- 0.2 # true value for the intercept
b1 <- 0.5 # true value for first beta
b2 <- 0.7 # true value for second beta
n <- 500 # sample size
X1 <- runif(n, -1, 1)
X2 <- runif(n, -1, 1)
Z <- b0 + b1 * X1 + b2 * X2
pr <- 1 / (1 + exp(-Z)) # inv logit function
Y <- rbinom(n, 1, pr)
data <- data.frame(cbind(X1, X2, Y))

## formatting the data for jags
datjags <- as.list(data)
datjags$N <- length(datjags$Y)

## creating jags model
model <- function()  {

  for(i in 1:N){
    Y[i] ~ dbern(p[i])  ## Bernoulli distribution of y_i
    logit(p[i]) <- mu[i]    ## Logit link function
    mu[i] <- b[1] +
      b[2] * X1[i] +
      b[3] * X2[i]
  }

  for(j in 1:3){
    b[j] ~ dnorm(0, 0.001) ## Use a coefficient vector for simplicity
  }

}

params <- c("b")
inits1 <- list("b" = rep(0, 3))
inits2 <- list("b" = rep(0, 3))
inits <- list(inits1, inits2)

## fitting the model with R2jags
set.seed(123)
fit <- R2jags::jags(data = datjags, inits = inits,
                    parameters.to.save = params, n.chains = 2, n.iter = 2000,
                    n.burnin = 1000, model.file = model)

## printing out table
object <- mcmcTab(fit, 
          ci = c(0.025, 0.975), 
          pars = NULL, 
          Pr = FALSE,
          ROPE = NULL)
object
}
}
